Fix: StartTimes & Offsets -> Floats
Former-commit-id: 0df596e2dea61062d6411c8772090ba36c26bae5 [formerly 6e77bf32e7e292703c52389a93fbeb5b38b5052b]
Former-commit-id: 079e3771309f580f7a4fd0f0697d88a4cc9906e3

+++ Assets/Scripts/Qua.Scripts/Classes/QuaParser.cs
@@ -161,7 +161,7 @@ namespace Qua.Scripts {
 
 						TimingPoint timing = new TimingPoint();
 
-						timing.StartTime = Int32.Parse(values[0]);
+						timing.StartTime = float.Parse(values[0]);
 						timing.BPM = float.Parse(values[1]);
 
 						quaFile.TimingPoints.Add(timing);
@@ -183,7 +183,7 @@ namespace Qua.Scripts {
 
 						SliderVelocity sv = new SliderVelocity();
 
-						sv.StartTime = Int32.Parse(values[0]);
+						sv.StartTime = float.Parse(values[0]);
 						sv.Multiplier = float.Parse(values[1]);
 						sv.Volume = Int32.Parse(values[2]);
